src-backend/requirements.txt
Flask
numpy
opencv-python-headless
Pillow
gunicorn
flask-cors
